Move signup, appointment css to assets folder
